Move SegmentEncodingSpec to avoid having everybody include the ChunkEncoder.;
